+++ src/main/java/net/earlystage/item/material/EarlyStageArmorMaterials.java
@@ -1,9 +1,6 @@
 package net.earlystage.item.material;
 
-import java.util.EnumMap;
-import java.util.List;
-import java.util.function.Supplier;
-
+import net.earlystage.EarlyStageMain;
 import net.earlystage.init.ItemInit;
 import net.minecraft.item.ArmorItem;
 import net.minecraft.item.ArmorMaterial;
@@ -13,12 +10,15 @@ import net.minecraft.registry.Registry;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
-import net.minecraft.util.Identifier;
 import net.minecraft.util.Util;
+
+import java.util.EnumMap;
+import java.util.List;
+import java.util.function.Supplier;
 
 public class EarlyStageArmorMaterials {
 
-    public static final RegistryEntry<ArmorMaterial> STEEL = register("earlystage:steel", Util.make(new EnumMap<>(ArmorItem.Type.class), map -> {
+    public static final RegistryEntry<ArmorMaterial> STEEL = register("steel", Util.make(new EnumMap<>(ArmorItem.Type.class), map -> {
         map.put(ArmorItem.Type.BOOTS, 2);
         map.put(ArmorItem.Type.LEGGINGS, 5);
         map.put(ArmorItem.Type.CHESTPLATE, 7);
@@ -26,15 +26,23 @@ public class EarlyStageArmorMaterials {
         map.put(ArmorItem.Type.BODY, 5);
     }), 9, SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0f, 0.0f, () -> Ingredient.ofItems(ItemInit.STEEL_INGOT));
 
+    public static final RegistryEntry<ArmorMaterial> DAMASCUS_STEEL = register("damascus_steel", Util.make(new EnumMap<>(ArmorItem.Type.class), map -> {
+        map.put(ArmorItem.Type.BOOTS, 3);
+        map.put(ArmorItem.Type.LEGGINGS, 6);
+        map.put(ArmorItem.Type.CHESTPLATE, 8);
+        map.put(ArmorItem.Type.HELMET, 3);
+        map.put(ArmorItem.Type.BODY, 5);
+    }), 9, SoundEvents.ITEM_ARMOR_EQUIP_IRON, 0.0f, 0.0f, () -> Ingredient.ofItems(ItemInit.DAMASCUS_STEEL_INGOT));
+
     private static RegistryEntry<ArmorMaterial> register(String id, EnumMap<ArmorItem.Type, Integer> defense, int enchantability, RegistryEntry<SoundEvent> equipSound, float toughness,
             float knockbackResistance, Supplier<Ingredient> repairIngredient) {
-        List<ArmorMaterial.Layer> list = List.of(new ArmorMaterial.Layer(Identifier.of(id)));
+        List<ArmorMaterial.Layer> list = List.of(new ArmorMaterial.Layer(EarlyStageMain.identifierOf(id)));
 
         EnumMap<ArmorItem.Type, Integer> enumMap = new EnumMap<ArmorItem.Type, Integer>(ArmorItem.Type.class);
         for (ArmorItem.Type type : ArmorItem.Type.values()) {
             enumMap.put(type, defense.get(type));
         }
-        return Registry.registerReference(Registries.ARMOR_MATERIAL, Identifier.of(id),
+        return Registry.registerReference(Registries.ARMOR_MATERIAL, EarlyStageMain.identifierOf(id),
                 new ArmorMaterial(enumMap, enchantability, equipSound, repairIngredient, list, toughness, knockbackResistance));
     }