+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ### Added:
-- 
+- Damascus steel set
 ### Fixed:
-- Fixed crash with blast furnace disabled slot
-- Fixed dehydration compat
+- 
 ### Changed:
 -